Move takePicture Composable into own file.
